requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
Bootstrap-Flask==2.4.0
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-bootstrap5==0.1.dev1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
